requirements.txt
# Core dependencies for semiconductor manufacturing self-learning module
crawl4ai>=0.3.0
langchain>=0.1.0
langchain-community>=0.0.20
chromadb>=0.4.0
sentence-transformers>=2.2.0
transformers>=4.30.0
torch>=2.0.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
faiss-cpu>=1.7.4
beautifulsoup4>=4.12.0
requests>=2.31.0
aiohttp>=3.8.0
asyncio-mqtt>=0.13.0
schedule>=1.2.0
python-dotenv>=1.0.0
pydantic>=2.0.0
fastapi>=0.100.0
uvicorn>=0.23.0
sqlalchemy>=2.0.0
alembic>=1.11.0
redis>=4.6.0
celery>=5.3.0
loguru>=0.7.0
rich>=13.0.0
typer>=0.9.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0
jupyter>=1.0.0
notebook>=7.0.0
ipykernel>=6.25.0
openai>=1.0.0
anthropic>=0.7.0
python-dateutil>=2.8.0
pytz>=2023.3
apscheduler>=3.10.0
streamlit>=1.28.0
psutil>=5.9.0
